# simulator/api/simulator/v1/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Simulator API
    description: The simulator service definition.
    version: 0.0.1
paths:
    /api/OnBetReq:
        post:
            tags:
                - Simulator
            operationId: Simulator_OnBetReq
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/BetReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BetRsp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/OnCancelBetReq:
        post:
            tags:
                - Simulator
            operationId: Simulator_OnCancelBetReq
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CancelBetReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CancelBetRsp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /api/OnGetBetListReq:
        post:
            tags:
                - Simulator
            operationId: Simulator_OnGetBetListReq
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/GetBetListReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetBetListRsp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /hello/{name}:
        get:
            tags:
                - Simulator
            description: Sends a greeting
            operationId: Simulator_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HelloRsp'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        BetReq:
            type: object
            properties:
                gameID:
                    type: integer
                    description: 游戏ID
                    format: int32
                uid:
                    type: integer
                    description: playerID
                    format: int64
        BetRsp:
            type: object
            properties: {}
        CancelBetReq:
            type: object
            properties: {}
        CancelBetRsp:
            type: object
            properties: {}
        GetBetListReq:
            type: object
            properties: {}
        GetBetListRsp:
            type: object
            properties: {}
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        HelloRsp:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
tags:
    - name: Simulator